requirements.txt
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
docker==7.0.0
future==0.18.3
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-telegram-bot==12.6.1
requests==2.31.0
six==1.16.0
tornado==6.4
urllib3==2.1.0
websocket-client==1.7.0
